requirements.txt
discord.py
python-dotenv
requests
psycopg2-binary
